This is just after working on the mini self project that got assigned in the tutorials. Worked on the Logger class.
The Logger class deals with logging the login time of the users. Where
this became a challenge was when I need to have the username move
script to script, so that I can use it to InputEntry with the Logger
class. I needed to use other classes inside of classes which caused me
to have to require those files in THAT class's script.l Just using the
initialize.php would not work since the other was wrong for the
Session class to call the User class and I could not rely on the
__autoload function since the one (& only) object (DatabseObject) does
not have same name as class. I could change this which will be on the
next commit.