This commit is really showing my success, as I have started to branch off from the tutorials. biggest thing is hashed_password.
So while working through the part to create, update and delete a user
I have found it makes more sence to me to do things a bit different
then the tutorials.
1. Was when making a new User I just want to pass in value to a
__construct method which cleans things up.

2. The challenge was making the necessary changes to go from a plain
text password to a fully hashed one. I had to focus on what I am
storing in the database, as well as what is stored in the instance of
the user. Difference being the plaint text password.

I had to figure out:
    - how to alter tables with mysql
    - how to find the user base off username to try and match the plain
    text pass to the stored hashed one. (this was big headache)