Worked through abstracting the attribute and values for mysql statements in the create and update function.
First I went through to the [table_names] for the mysql and changed it
so that it does late static binding and can be place in the parent
class now.

However, because I have gone ahead and added in password encryption,
I have decided to override the create and update functions in the User
class. This also gives me a way to see the two different methods of
doing this.

database_object.php got the generic versions of the methods so that
I can have child classes for photos and comments later on.