I was able to fix the re-writing of the create() and update() methods since I now have a static variable that is an array which contains the column headers of the database tables.
Before I had to re-write the methods so that I was able to work only
with the database fields since the attribute() method in the
DatabaseObject was returning every attribute.

At fist I thought I did not need to include the 'id' column however
this became a problem when I went from database to User object like in
login. The instantiate() method will not know the 'id' of the User
and therefore not be able to set the session 'id' to confirm login.

Also with regards to encrypting the password I have set up so that I
have to run a method and pass is a new password which would come from
a form post to be encrypted and then run update() to update the
database itself