Have started the creation of the Photograph class.
The class attributes were set and explained with comments. These are
just from the uploaded file associative array which in the super
global array $_FILE.

Work on the method to set take in that associative and set the value
to the attributes of the class after checking from php errors as well
as if the file is even there to begin with.

Side note: had to modify the directory so that the web server can read,
write and execute the directory that contains the images.

Also created the database table in mysql.