Done making the Photograph class.
Have the ability to upload photos and list them on the website now.
Without watching more I am sure I can make the page for display the
larger size of each photo.

There were functions added to make the size a better display as well
as dynamic way of setting the img source in the list display part.