Added in the comment table to mysql, as well as the Comment object itself.
Since working with the created time for the comments I was thinking of
adding similar actions to the User and Photograph classes as well...
This however requires that I change the tables in mysql... can be done
just not getting into that at the moment.

I did change the name of the log_object.php file so that it will match
the name of the logger class. This is to fix the minor bug I had
awhile ago about the ordering of the initialize.php script and classes
calling other classes.