I added in the form to create a new user as well as a page to view the list of users and a link to delete Users.
However I am not sure what to do about requiring the user's password
before deleting. I would like maybe a popup javascript window. but I
do not know yet if I can then pass that to the php function in the
class User. In this case $user->delete('pass');

The listing of the users is the same right now as the photos which I
do no think is the best looking.