Worked on a page to create a User, delete the user as well as in the public side a form page to submit comments to photos as well to view the comments in the admin area so that they can be deleted
I made a simple javascript confirm since I was unable to figure out
how to prompt for the password and transfer that variable to the
delete_user.php page. This script reference is found in the admin
footer layout page.

The listing of users can be worked on to have a better layout then the
same as the photos.